requirements.txt
# Smart Agriculture System - Dependencies
# Install with: pip install -r requirements.txt

streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
Pillow>=10.0.0
plotly>=5.15.0
requests>=2.25.0
google-generativeai>=0.3.0
